requirements.txt
# Weather Proxy API dependencies
fastapi==0.95.1
uvicorn[standard]==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
pydantic==1.10.7
redis==4.5.5
pytest==8.2.1
pytest-asyncio==0.23.5